+++ Library/_TEST/_TEST_treap/_TEST_UQF_treap/a.cpp
@@ -139,7 +139,7 @@ struct modular_fixed_base{
 	modular_fixed_base operator--(int){ modular_fixed_base result(*this); *this += _mod - 1; return result; }
 	modular_fixed_base operator-() const{ return modular_fixed_base(_mod - data); }
 	modular_fixed_base &operator*=(const modular_fixed_base &rhs){
-		data = (unsigned long long)data * rhs.data % _mod;
+		data = (uint64_t)data * rhs.data % _mod;
 		return *this;
 	}
 	template<class T, typename enable_if<is_integral<T>::value>::type* = nullptr>
@@ -191,7 +191,7 @@ template<uint _mod> modular_fixed_base<_mod> operator/(const modular_fixed_base<
 template<uint _mod, class T, typename enable_if<is_integral<T>::value>::type* = nullptr> modular_fixed_base<_mod> operator/(const modular_fixed_base<_mod> &lhs, T rhs) { return modular_fixed_base<_mod>(lhs) /= rhs; }
 template<uint _mod, class T, typename enable_if<is_integral<T>::value>::type* = nullptr> modular_fixed_base<_mod> operator/(T lhs, const modular_fixed_base<_mod> &rhs) { return modular_fixed_base<_mod>(lhs) /= rhs; }
 template<uint _mod> istream &operator>>(istream &in, modular_fixed_base<_mod> &number){
-	long long x;
+	int64_t x;
 	in >> x;
 	number.data = modular_fixed_base<_mod>::normalize(x);
 	return in;
